Leo/improvement

+++ cmd/nutmix/htlc_route_test.go
@@ -319,7 +319,7 @@ func makeHTLCSpendCondition(preimage string, nSigs uint, pubkeys []*secp256k1.Pu
 	spendCondition.Data.Data = hex.EncodeToString(parsedPreimage[:])
 	spendCondition.Data.Tags.Pubkeys = pubkeys
 	spendCondition.Data.Tags.NSigs = nSigs
-	spendCondition.Data.Tags.Locktime = locktime
+	spendCondition.Data.Tags.Locktime = uint64(locktime)
 	spendCondition.Data.Tags.Sigflag = sigflag
 	spendCondition.Data.Tags.Refund = refundPubkey
